+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.2] - 2026-04-21
+
+### Added
+
+- `uk_input`: add `margin` to apply UIkit's `uk-margin` spacing to generated
+  wrappers for non-hidden inputs.
+
 ## [0.7.1] - 2026-04-20
 
 ### Documentation